svidetelBot/requirements.txt
telebot
requests
numpy
python-dotenv==1.1.1